+++ R/partial_dep.R
@@ -33,12 +33,13 @@
 #' @inheritParams multivariate_grid
 #' @param grid Evaluation grid. A vector (if `length(v) == 1L`), or a matrix/data.frame 
 #'   otherwise. If `NULL`, calculated via [multivariate_grid()].
-#' @param BY Optional grouping vector or a column name. The partial dependence
+#' @param BY Optional grouping vector or column name. The partial dependence
 #'   function is calculated per `BY` group. Each `BY` group
 #'   uses the same evaluation grid to improve assessment of (non-)additivity.
 #'   Numeric `BY` variables with more than `by_size` disjoint values will be 
-#'   binned into `by_size` quantile groups of similar size. Subsampling of `X` is done
-#'   within group to improve robustness.
+#'   binned into `by_size` quantile groups of similar size. To improve robustness,
+#'   subsampling of `X` is done within group. This only applies to `BY` groups with
+#'   more than `n_max` rows.
 #' @param by_size Numeric `BY` variables with more than `by_size` unique values will
 #'   be binned into quantile groups. Only relevant if `BY` is not `NULL`.
 #' @returns 
@@ -78,12 +79,14 @@
 #' plot(partial_dep(fit, v = c("Species", "Petal.Width"), X = iris), rotate_x = TRUE)
 #' 
 #' # Multivariate, multivariable, and BY (no plot available)
-#' pd <- partial_dep(fit, v = c("Petal.Width", "Petal.Length"), X = iris, BY = "Species")
+#' pd <- partial_dep(
+#'   fit, v = c("Petal.Width", "Petal.Length"), X = iris, BY = "Species"
+#' )
 #' pd
 #' 
 #' # MODEL 3: Gamma GLM -> pass options to predict() via ...
 #' fit <- glm(Sepal.Length ~ ., data = iris, family = Gamma(link = log))
-#' plot(partial_dep(fit, v = "Petal.Length", X = iris))
+#' plot(partial_dep(fit, v = "Petal.Length", X = iris), show_points = FALSE)
 #' plot(partial_dep(fit, v = "Petal.Length", X = iris, type = "response"))
 partial_dep <- function(object, ...) {
   UseMethod("partial_dep")
@@ -109,45 +112,28 @@ partial_dep.default <- function(object, v, X, pred_fun = stats::predict,
   
   # The function itself is called per BY group
   if (!is.null(BY)) {
-    if (length(BY) == 1L && BY %in% colnames(X)) {
-      by_name <- BY
-      BY <- X[, by_name]
-    } else {
-      stopifnot(
-        NCOL(BY) == 1L,
-        is.vector(BY) || is.factor(BY),
-        length(BY) == nrow(X)
-      )
-      by_name <- "Group"
-    }
-    
-    # Binning
-    by_values <- unique(BY)
-    if (is.numeric(BY) && length(by_values) > by_size) {
-      BY <- qcut(BY, m = by_size)
-      by_values <- unique(BY)
-    }
-    
-    pd_list <- vector("list", length = length(by_values))
-    for (i in seq_along(by_values)) {
-      b <- by_values[i]
+    BY2 <- prepare_by(BY = BY, X = X, by_size = by_size)
+    mm <- length(BY2$by_values)
+    pd_list <- vector("list", length = mm)
+    for (i in seq_len(mm)) {
+      b <- BY2$by_values[i]
       out <- partial_dep.default(
         object = object, 
         v = v, 
-        X = X[BY %in% b, , drop = FALSE], 
+        X = X[BY2$BY %in% b, , drop = FALSE], 
         pred_fun = pred_fun,
         grid = grid,
         n_max = n_max, 
-        w = if (!is.null(w)) w[BY %in% b],
+        w = if (!is.null(w)) w[BY2$BY %in% b],
         ...
       )
       pd_list[[i]] <- out[["data"]]
     }
     pd <- do.call(rbind, c(pd_list, list(make.row.names = FALSE)))
-    BY_rep <- rep(by_values, each = NROW(grid))
-    BY_rep <- stats::setNames(as.data.frame(BY_rep), by_name)
+    BY_rep <- rep(BY2$by_values, each = NROW(grid))
+    BY_rep <- stats::setNames(as.data.frame(BY_rep), BY2$by_name)
     out[["data"]] <- cbind.data.frame(BY_rep, pd)
-    out[["by_name"]] <- by_name
+    out[["by_name"]] <- BY2$by_name
     
     return(structure(out, class = "partial_dep"))
   }
@@ -287,6 +273,7 @@ print.partial_dep <- function(x, n = 3L, ...) {
 #' @param color Color of lines and points (in case there is no color/fill aesthetic).
 #'   The default equals the global option `hstats.color = "#3b528b"`. 
 #'   To change the global option, use `options(stats.color = new value)`.
+#' @param multi_output How should multi-output models (with single `v`) be represented? 
 #' @param show_points Logical flag indicating whether to show points (default) or not.
 #' @param ... Arguments passed to geometries.
 #' @inheritParams plot.hstats_matrix
@@ -295,9 +282,14 @@ print.partial_dep <- function(x, n = 3L, ...) {
 #' @seealso See [partial_dep()] for examples.
 plot.partial_dep <- function(x,
                              color = getOption("hstats.color"),
-                             facet_scales = "free_y", 
-                             rotate_x = FALSE,
-                             show_points = TRUE, ...) {
+                             multi_output = c("grouped", "facets"),
+                             viridis_args = getOption("hstats.viridis_args"),
+                             facet_scales = "free_y",
+                             ncol = 2L, rotate_x = FALSE, show_points = TRUE, ...) {
+  multi_output <- match.arg(multi_output)
+  if (is.null(viridis_args)) {
+    viridis_args <- list()
+  }
   v <- x[["v"]]
   by_name <- x[["by_name"]]
   K <- x[["K"]]
@@ -332,19 +324,20 @@ plot.partial_dep <- function(x,
       }
     }
     if (K > 1L) {
-      p <- p + ggplot2::facet_wrap(~ varying_, scales = facet_scales)
+      p <- p + ggplot2::facet_wrap(~ varying_, ncol = ncol, scales = facet_scales)
     }
   } else if (length(v) == 2L) {
     # Heat maps
     p <- ggplot2::ggplot(
       data, ggplot2::aes(x = .data[[v[1L]]], y = .data[[v[2L]]], fill = value_)
     ) + 
+      do.call(ggplot2::scale_fill_viridis_c, viridis_args) +
       ggplot2::geom_tile(...) +
       ggplot2::labs(fill = "PD")
     
     if (K > 1L || !is.null(by_name)) {  # Only one is possible
       wrp <- if (K > 1L) "varying_" else by_name
-      p <- p + ggplot2::facet_wrap(wrp, scales = facet_scales)
+      p <- p + ggplot2::facet_wrap(wrp, ncol = ncol, scales = facet_scales)
     }
   }
   if (rotate_x) p + rotate_x_labs() else p