+++ R/perm_importance.R
@@ -7,7 +7,7 @@
 #' loss when shuffling the corresponding feature values before calculating predictions.
 #' By default, the process is repeated `m_rep = 4` times, and the results are averaged.
 #' In most of the cases, importance values should be derived from an independent test
-#' data set.
+#' data set. Set `normalize = TRUE` to get *relative* increases in average loss.
 #' 
 #' @inheritSection average_loss Losses
 #'