+++ R/statistics.R
@@ -138,10 +138,12 @@ print.hstats_matrix <- function(x, top_m = Inf, ...) {
 #'   `options(stats.fill = new value)`.
 #' @param multi_output How should multi-output models be represented? 
 #'   Either as "grouped" barplot (the default) or via "facets".
-#' @param scale_fill_d Discrete fill scale for grouped bars. The default equals the 
-#'   global option `hstats.scale_fill_d`, which equals 
-#'   `scale_fill_viridis_d(begin = 0.25, end = 0.85, option = "inferno")`. 
-#'   To change the global option, use `options(hstats.scale_fill_d = new value)`.
+#' @param viridis_args List of viridis color scale arguments, see
+#'   `[ggplot2::scale_color_viridis_c()]`. 
+#'   The default points to the global option `hstats.viridis_args`, 
+#'   which corresponds to `list(begin = 0.25, end = 0.85, option = "B")`.
+#'   E.g., to switch to a standard viridis scale, you can change the default via 
+#'   `options(hstats.viridis_args = list())`, or set `viridis_args = list()`. 
 #' @param facet_scales Value passed as `scales` argument to `[ggplot2::facet_wrap()]`.
 #' @param ncol Passed to `[ggplot2::facet_wrap()]`.
 #' @param rotate_x Should x axis labels be rotated by 45 degrees?
@@ -154,12 +156,15 @@ print.hstats_matrix <- function(x, top_m = Inf, ...) {
 plot.hstats_matrix <- function(x, top_m = 15L,
                                fill = getOption("hstats.fill"),
                                multi_output = c("grouped", "facets"),
-                               scale_fill_d = getOption("hstats.scale_fill_d"),
+                               viridis_args = getOption("hstats.viridis_args"),
                                facet_scales = "free", 
                                ncol = 2L, rotate_x = FALSE,
                                err_type = c("SE", "SD", "No"), ...) {
   err_type <- match.arg(err_type)
   multi_output <- match.arg(multi_output)
+  if (is.null(viridis_args)) {
+    viridis_args <- list()
+  }
   
   M <- x[["M"]]
   if (is.null(M)) {
@@ -192,7 +197,7 @@ plot.hstats_matrix <- function(x, top_m = 15L,
       ggplot2::aes(fill = varying_), stat = "identity", position = "dodge", ...
     ) + 
       ggplot2::theme(legend.title = ggplot2::element_blank()) +
-      scale_fill_d
+      do.call(ggplot2::scale_fill_viridis_d, viridis_args)
   }
   if (err_type != "No") {
     if (!grouped) {