+++ NEWS.md
@@ -4,18 +4,20 @@ This release mainly changes the *output*. The numeric results are unchanged.
 
 ## Major changes
 
-- Revised plots: The colors have changed and can (also) be controlled via global options. For instance, to change the fill color of all bars, set `options(stats.fill = new value)`. Value labels are more clear, and there are more options. Stacked barplots use viridis.
+- Revised plots: The colors and color palettes have changed and can (also) be controlled via global options. For instance, to change the fill color of all bars, set `options(hstats.fill = new value)`. Value labels are more clear, and there are more options. Varying color/fill scales now use viridis (inferno). This can be modified on the fly or via `options(hstats.viridis_args = list(...))`.
 - "hstats_matrix" object: All statistics functions, e.g., `h2_pairwise()` or `perm_importance()`, return a "hstats_matrix" object. The values are stored in `$M` and can be plotted via `plot()`.
 - `perm_importance()`: The `perms` argument has been changed to `m_rep`.
 - All `print()` and `summary()` methods have been revised.
 
 ## Minor changes
 
-- Statistics: Their argument `top_m` has been moved to the `plot()` method.
+- Statistics: The argument `top_m` has been moved to the `plot()` method.
 - Statistics: The clipping threshold `eps` of squared numerator statistics has been reduced from `1e-8` to `1e-10`. It is now handled in `hstats()` instead of the statistic functions.
 - `H-squared`: The $H^2$ statistic stored in a "hstats" object is now a matrix with one row (it was a vector).
 - `pd_importance()`: The "hstats" object now contains pre-calculated PD-based importance values in `$pd_importance`.
 - `summary.hstats()` now returns an object of class "hstats_summary" instead of "summary_hstats".
+- `average_loss()` is more flexible regarding the group `BY` argument. It can also be a variable *name*. Non-discrete `BY` variables are now automatically binned. Like `partial_dep()`, binning is controlled by the `by_size = 4` argument.
+- `average_loss()` also returns a "hstats_matrix" object with `print()` and `plot()` method. The values can be extracted via `$M`.
 
 # hstats 0.3.0